+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 > 2023-01-16
 
--   update `@fullpage/react-fullpage` to version 0.1.26
+-   remove `@fullpage/react-fullpage` since has snyk license issue (about page not used yet)
+-   upgrade react-pdf (brochure page), react-select (in axa page not used yet), react-responsive-carousel (download page and promo page) , and next version
 
 ### [v1.10.1]